refactor: 폴더 구조 mvc로 변경
