+++ src/test/java/lotto/LottoMarketTest.java
@@ -1,5 +1,8 @@
 package lotto;
 
+import lotto.controller.LottoMarket;
+import lotto.model.Lottos;
+import lotto.model.RandomLottoGenerator;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
